+++ pyam/utils.py
@@ -353,8 +353,10 @@ def _escape_regexp(s):
 
 def years_match(data, years):
     """Return rows where data matches year"""
-    years = [years] if isinstance(years, int) else years
-    dt = datetime.datetime
+    years = [years] if (
+        isinstance(years, (int, np.int64))
+    ) else years
+    dt = (datetime.datetime, np.datetime64)
     if isinstance(years, dt) or isinstance(years[0], dt):
         error_msg = "`year` can only be filtered with ints or lists of ints"
         raise TypeError(error_msg)
@@ -423,9 +425,11 @@ def time_match(data, times, conv_codes, strptime_attr, name):
 def datetime_match(data, dts):
     """Matching of datetimes in time columns for data filtering"""
     dts = dts if islistable(dts) else [dts]
-    if any([not isinstance(i, datetime.datetime) for i in dts]):
+    if any([not (
+            isinstance(i, (datetime.datetime, np.datetime64))
+    ) for i in dts]):
         error_msg = (
-            "`time` can only be filtered by datetimes"
+            "`time` can only be filtered by datetimes and datetime64s"
         )
         raise TypeError(error_msg)
     return data.isin(dts)